Refactor profile and activity screens; rename EditProfileScreen, update navigation to Message screen, and improve user profile handling in ActivityCard. Remove unused ChatDetailScreen and OrganizerProfileScreen components, and enhance MessageScreenHeader with bio display.
