+++ sportsphere-app/components/TabNavigator.jsx
@@ -7,13 +7,8 @@ import ChatScreen from '../screens/ChatScreen';
 import MapScreen from '../screens/MapScreen';
 import ProfileScreen from '../screens/ProfileScreen';
 import { COLORS, ROUNDED, SPACING, SHADOW, SIZE } from '../global';
-import ActivityScreenHeader from './ActivityScreenHeader';
-import TitleScreenHeader from './TitleScreenHeader';
-import ProfileScreenHeader from './ProfileScreenHeader';
 import ActivityStack from './ActivityStack';
 import { useState } from 'react';
-import OrganizerProfileScreenHeader from './OrganizerProfileScreenHeader';
-import OrganizerProfileScreen from '../screens/OrganizerProfileScreen';
 import LocationManager from './LocationManager';
 import { UserContext } from '../context/UserProvider';
 import { QueryContext } from '../context/QueryProvider';
